Add files via upload
A text to QR code generator is a software or online tool that converts plain text or alphanumeric data into a Quick Response (QR) code. QR codes are two-dimensional barcodes that can store various types of information, such as text, URLs, contact information, Wi-Fi credentials, and more. The primary purpose of a text to QR code generator is to encode the provided text as a QR code, which can then be easily scanned and decoded using a QR code reader or a smartphone camera